feat: Add constanstElement.js and bookClass.js
