+++ modules/constantElement.js
@@ -1,0 +1,11 @@
+// Select the Elements
+export const container = document.querySelector('.books-container');
+export const addBtn = document.querySelector('form');
+export const titleInput = document.querySelector('.title-input');
+export const authorInput = document.querySelector('.author-input');
+export const errorMsg = document.querySelector('.error');
+export const successMsg = document.querySelector('.success');
+export const listSection = document.querySelector('.books-section');
+export const addNewSection = document.querySelector('.data-entry-section');
+export const contactSection = document.querySelector('.contact-section');
+export const [listNav, addNewNav, contactNav] = document.querySelectorAll('.list-item');
